Enhance cross-compilation support for FastFinder; add ARM64 and i386 builds for Linux and Windows, update Dockerfiles, helper scripts, and Makefile for new binaries

+++ yaraprocessing.go
@@ -234,7 +234,7 @@ func FileAnalyzeYaraMatch(path string, rules *yara.Rules, maxFileSizeScan int, c
 	}
 
 	// cancel analysis if file size is greater than 2Gb
-	if len(content) > 1024*1024*2048 {
+	if int64(len(content)) > 1024*1024*2048 {
 		LogMessage(LOG_ERROR, fmt.Sprintf("File size is greater than %dMb, skipping", maxFileSizeScan), path)
 		return false
 	}
